requirements.txt
bitarray==2.5.1
einops==0.8.1
jsonpickle==2.2.0
lempel_ziv_complexity==0.2.2
matplotlib==3.8.3
matplotlib==3.7.3
moviepy==2.2.1
numba==0.60.0
numpy==2.3.2
pandas==2.0.3
Pillow==11.3.0
plotly==5.14.0
scipy==1.16.1
seaborn==0.11.0
tvb_framework==2.9
tvb_library==2.9
tvb_storage==2.9
